test(bpmn): hoist shared model builders into tests/bpmn_models.py
